Crate score

These contain the server “cores”. These are able to startup the server (bootstrap) to a running state and then execute tasks. This is where modules are logically ordered based on their depenedncies for execution. Some of these are task-only i.e. reindexing, and some of these launch the server into a fully operational state (https, ldap, etc).

Generally, this is the “entry point” where the server begins to run, and the entry point for all client traffic which is then directed to the various actors.

Functions

backup_server_core

create_server_core

dbscan_get_id2entry_core

dbscan_list_id2entry_core

dbscan_list_index_analysis_core

dbscan_list_index_core

dbscan_list_indexes_core

domain_rename_core

recover_account_core

reindex_server_core

restore_server_core

vacuum_server_core

verify_server_core